repo/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic>=2.11.3,<3
psycopg[binary]==3.2.13
fpdf2==2.8.3
pytest==8.3.3
httpx==0.27.2
